% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttheme.R
\name{add_ttheme}
\alias{add_ttheme}
\title{Modify the properties of the ttheme}
\usage{
add_ttheme(t1, t2, t2name)
}
\arguments{
\item{t1}{The existing ttheme object.}

\item{t2}{The ttheme object to add to `t1`.}

\item{t2name}{A name of the t2 object. This is used for printing informative
error messages.}
}
\description{
Modify the properties of the ttheme
}
