% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcols.R
\name{default_tcol}
\alias{default_tcol}
\title{Create a default column for a ggtable}
\usage{
default_tcol(var, is_numeric)
}
\arguments{
\item{var}{A character string to specify the variable with which to create a
default column.}

\item{is_numeric}{A logical to indicate if the variable is numeric.}
}
\description{
Create a default column for a ggtable
}
