% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcols.R
\name{default_tcol_numeric}
\alias{default_tcol_numeric}
\title{Create a default numeric column for a ggtable}
\usage{
default_tcol_numeric(var)
}
\arguments{
\item{var}{A character string to specify the variable with which to create a
default column.}
}
\description{
Create a default numeric column for a ggtable
}
