% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcols.R
\name{default_tcols}
\alias{default_tcols}
\title{Create the default columns for a ggtable}
\usage{
default_tcols(data)
}
\arguments{
\item{data}{The data with which to create default columns.}
}
\description{
Create the default columns for a ggtable
}
