% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtable.R
\name{ggtable}
\alias{ggtable}
\title{Create a new ggtable}
\usage{
ggtable(data)
}
\arguments{
\item{data}{Data to display on the table.}
}
\description{
`ggtable` initializes a ggtable object.  Use the `+` operator to add
specifications and features to the table.
}
