% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtable-build.R
\name{ggtable_build}
\alias{ggtable_build}
\title{Build a ggtable for rendering}
\usage{
ggtable_build(table)
}
\arguments{
\item{table}{A ggtable object to build.}
}
\description{
Build a ggtable for rendering
}
