% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtable-build.R
\name{ggtable_gtable}
\alias{ggtable_gtable}
\title{Create a table from a built ggtable object}
\usage{
ggtable_gtable(built)
}
\arguments{
\item{built}{A built ggtable with which to create a ggtable.}
}
\description{
Create a table from a built ggtable object
}
