% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{ggtsave}
\alias{ggtsave}
\title{PDF a ggtable}
\usage{
ggtsave(x, filename, width = 11, height = 8.5)
}
\arguments{
\item{table}{A table to save.}

\item{path}{A location to save the PDF.}
}
\description{
PDF a ggtable
}
