% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtable-construction.R
\name{+.ggt}
\alias{+.ggt}
\title{Add features to a ggtable}
\usage{
\method{+}{ggt}(e1, e2)
}
\arguments{
\item{e1}{An object of class `ggtable` or a `ttheme`.}

\item{e2}{A ggtable component, including `tcols`, `ttheme`, `tlabs`,
`tbreaks`, `tpanels`, and `theaders`.}
}
\description{
Add features to a ggtable
}
