% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbreaks.R
\name{tbreaks}
\alias{tbreaks}
\title{Add page breaks to a ggtable}
\usage{
tbreaks(breaks)
}
\arguments{
\item{breaks}{A vector which specifies the last row number that will appear
on each page.}
}
\description{
Add page breaks to a ggtable
}
