% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcols.R
\name{tcol}
\alias{tcol}
\title{Create a column for a ggtable}
\usage{
tcol(var, header, format_fun, width, hjust, x, header_hjust, header_x, wrap_len)
}
\arguments{
\item{var}{A character string to specify the variable to displayed on the
table.}

\item{header}{A character string to specify the column header to display on
the table.}

\item{format_fun}{A function to apply to the values in the column that will
generate the values to display on the table.}

\item{width}{A unit value (i.e., unit(2, "cm")) to specificy the width of the
column.}

\item{hjust}{A numeric value between 0 (left) and 1 (right) to specify the
horizontal justification for the data values.}

\item{x}{A numeric value between 0 (left) and 1 (right) to specify the
horizontal position for the data values.}

\item{header_hjust}{A numeric value between 0 (left) and 1 (right) to specify
the horizontal justification for the headers.}

\item{header_x}{A numeric value between 0 (left) and 1 (right) to specify the
horizontal position for the headers.}

\item{wrap_len}{the target length at which the values in the column should
wrap.}
}
\description{
Create a column for a ggtable
}
