% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcols.R
\name{tcol_blank}
\alias{tcol_blank}
\title{Create a spacer column}
\usage{
tcol_blank(width = grid::unit(1, "cm"))
}
\arguments{
\item{width}{A unit value (i.e., unit(2, "cm")) to specificy the width of the
column.}
}
\description{
Create a spacer column
}
