% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcols.R
\name{tcols}
\alias{tcols}
\title{Modify the columns of a ggtable}
\usage{
tcols(...)
}
\arguments{
\item{...}{tcols or a list of tcols with which to update your table.}
}
\description{
Modify the columns of a ggtable
}
