% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theaders.R
\name{theaders}
\alias{theaders}
\title{Add super headers to a ggtable}
\usage{
theaders(...)
}
\arguments{
\item{...}{Vectors or a list of vectors which specify the columns that each
superheader includes.  The name of each argument specifies the title of the
superheader.}
}
\description{
Add super headers to a ggtable
}
