% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlabs.R
\name{tlabs}
\alias{tlabs}
\title{Modify the title and caption of a ggtable}
\usage{
tlabs(..., wrap_note = TRUE, wrap_source = TRUE, wrap_len = 80,
  source_exdent = 0, note_exdent = 6)
}
\arguments{
\item{wrap_note}{A logical value to indicate if the note should wrap.}

\item{wrap_source}{A logical value to indicate if the source should wrap.}

\item{wrap_len}{An integer indicating the target wrap length.}

\item{source_exdent}{An integer indicating the indentation of subsequent
lines in the source.}

\item{note_exdent}{An integer indicating the indentation of subsequent lines
in the note.}

\item{title}{The title for the table.}

\item{subtitle}{The subtitle for the table.}

\item{source}{The source line for the table.}

\item{note}{The note section for the table.}
}
\description{
Modify the title and caption of a ggtable
}
