% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpanels.R
\name{tpanels}
\alias{tpanels}
\title{Divide the ggtable into panels}
\usage{
tpanels(var)
}
\arguments{
\item{var}{A variable by which to divide the ggtable into panels.  The values
of the variable specify the panel titles.}
}
\description{
Divide the ggtable into panels
}
