% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttheme.R
\name{ttheme}
\alias{ttheme}
\title{Modify the components of a ttheme}
\usage{
ttheme(text, rect, title, table.title, table.subtitle, panel.title, caption,
  caption.source, caption.note, header, header.main, header.super,
  header.background, data, data.background, margin)
}
\arguments{
\item{text}{All text elements (`element_text`).}

\item{rect}{All rectangle elements (`element_rect`).}

\item{title}{All title elements including the title and subtitle
(`element_text`; inherits from `text`).}

\item{table.title}{The table title (`element_text`; inherits from `title`).}

\item{table.subtitle}{The table subtitle (`element_text`; inherits from
`title`).}

\item{panel.title}{The title for each panel (`element_text`; inherits from
`text`).}

\item{caption}{All caption elements including the source and the note
(`element_text`; inherits from `text`).}

\item{caption.source}{The source line for the table (`element_text`; inherits
from `caption`).}

\item{caption.note}{The note for the table (`element_text`; inherits from
`caption`).}

\item{header}{The column headers (`element_text`; inherits from `text`).}

\item{header.main}{The column headers (`element_text`; inherits from `header`).}

\item{header.super}{The column super headers (`element_text`; inherits from `header`).}

\item{header.background}{The background for the column headers
(`element_rect`; inherits from `rect`).}

\item{data}{The data contents of the table (`element_text`; inherits from
`text`).}

\item{data.background}{The background for the data contents of the table
(`element_rect`; inherits from `rect`).}

\item{margin}{The margins around the entire table (`unit` with the sizes of
the top, right, bottom, and left margins).}
}
\description{
Modify the components of a ttheme
}
