% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttheme-elements.R
\name{ttheme_calc_element}
\alias{ttheme_calc_element}
\title{Calculate element properties by inheriting properties from its parents}
\usage{
ttheme_calc_element(element, ttheme)
}
\arguments{
\item{element}{The name of the ttheme element to calculate.}

\item{ttheme}{A ttheme object.}
}
\description{
Calculate element properties by inheriting properties from its parents
}
