% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttheme_cr.R
\name{ttheme_cr}
\alias{ttheme_cr}
\title{The default theme for ggtable}
\usage{
ttheme_cr(base_size = base_font$size, font_family = base_font$family)
}
\arguments{
\item{base_size}{The base font size.}

\item{font_family}{The base font family.}
}
\description{
The default theme for ggtable
}
