% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttheme.R
\name{ttheme_helper}
\alias{ttheme_helper}
\title{Process a ttheme to use with `tableGrob`}
\usage{
ttheme_helper(ttheme, tcols)
}
\arguments{
\item{ttheme}{A ttheme to process.}
}
\description{
Process a ttheme to use with `tableGrob`
}
