% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbreaks.R
\name{update_tbreaks}
\alias{update_tbreaks}
\title{Update the tbreaks for a ggtable}
\usage{
update_tbreaks(table, new_tbreaks)
}
\arguments{
\item{table}{A table to update.}

\item{new_tbreaks}{The new tbreaks object.}
}
\description{
Update the tbreaks for a ggtable
}
