% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcols.R
\name{update_tcols}
\alias{update_tcols}
\title{Update the columns on a ggtable}
\usage{
update_tcols(table, new_tcols)
}
\arguments{
\item{table}{A table to update.}

\item{tcols}{The new columns.}
}
\description{
Update the columns on a ggtable
}
