% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theaders.R
\name{update_theaders}
\alias{update_theaders}
\title{Update the super headers on a ggtable}
\usage{
update_theaders(table, new_theaders)
}
\arguments{
\item{table}{A table to update.}

\item{newtheaders}{The new superheaders.}
}
\description{
Update the super headers on a ggtable
}
