% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlabs.R
\name{update_tlabs}
\alias{update_tlabs}
\title{Update the labels on a ggtable}
\usage{
update_tlabs(table, tlabs)
}
\arguments{
\item{table}{A table to modify.}

\item{tlabs}{The new labels.}
}
\description{
Update the labels on a ggtable
}
