% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpanels.R
\name{update_tpanels}
\alias{update_tpanels}
\title{Update the tpanels for a ggtable}
\usage{
update_tpanels(table, new_tpanels)
}
\arguments{
\item{table}{A table to update.}

\item{new_tpanels}{The new tpanels object.}
}
\description{
Update the tpanels for a ggtable
}
