% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttheme.R
\name{update_ttheme}
\alias{update_ttheme}
\title{Update the ttheme of a ggtable}
\usage{
update_ttheme(table, new_ttheme)
}
\arguments{
\item{table}{A table to modify.}

\item{ttheme}{A new ttheme to add to the existing ttheme.}
}
\description{
Update the ttheme of a ggtable
}
